requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
pillow
flask
flask-cors
python-dotenv
tqdm